BigDecimal conversion and primitive PIN checking: - Replaced doubles with BigDecimal when it came to money to enforce best practices.    - Referenced it inside of the test cases    - Reworked the api layer to make it work - Enforced some more visual queues on the api layer. - Added changes to the slides - Added SQL Error handling and logging
Todo:
- Verify that everything still works and fix whatever does not
- Maybe add other exception types
- Maybe actually change PINs to Strings

+++ src/main/java/com/revature/persistence/AccountDAO.java
@@ -1,5 +1,6 @@
 package com.revature.persistence;
 
+import java.math.BigDecimal;
 import java.sql.Connection;
 
 import com.revature.domain.Account;
@@ -11,7 +12,7 @@ public interface AccountDAO {
     Account login(int accountId, int pin);
 
     // This is only called after a deposit/withdraw and thus can accept a connection
-    Account updateBalance(int accountId, double amount, Connection connection);
+    Account updateBalance(int accountId, BigDecimal amount, Connection connection);
 
     void updatePin(int accountId, int newPin);
